Docs(Dgraph): Update "More about Dgraph Zero" to reflect the output of the /state endpoint, and provide an edit pass (#6769)
* Update dgraph-zero.md

Added new /state sample syntax, updated notes on what the syntax shows, started an edit pass for style.

* Update dgraph-zero.md

Ready for review.

From reading this, it sounds like /assign and /removeNode are POST endpoints, but they are listed in the doc as GET endpoints. Need to confirm with reviewers.

* Update dgraph-zero.md

Added headings and changed heading levels.

* Update wiki/content/deploy/dgraph-zero.md

Co-authored-by: Damián Parrino <bucanero@users.noreply.github.com>

* Update dgraph-zero.md

Incorporated review feedback by Damian and Naman

* Merge branch 'dgraph-zero' of https://github.com/dgraph-io/dgraph into dgraph-zero

* Update dgraph-zero.md

Formatting tweak

* Update dgraph-zero.md

more formatting tweaks to fix bad site rendering.

* Update dgraph-zero.md

Fixed annoying highlighting of the word "for" when the error message was formatted as ```bash by formatting it as ```nix instead (generic command-ilne formatting)

* Update dgraph-zero.md

broke Group 1 and Group 0 members apart in this list, and re-ordered list contents to match the order of nodes in the example JSON.