[Docs revamp] consolidated all export related tasks in one section of howtos (#355)
Description: Please include the following in your PR description:
* Consolidated all export related information and rewrote the content in
the how to framework.

Note: Create and edit docs in the `main` branch when you can, so that we
only cherry-pick out of `main`, not into `main`.
-->
_ i have remved references to check queue in the export content because
as Rahul confirmed "Yes, that's correct. We cannot query queued/running
tasks on any v21.03.1 (and beyond) EE or OSS builds. (edited) "


Fixes: https://github.com/dgraph-io/dgraph-docs/issues/352
https://github.com/dgraph-io/dgraph-docs/issues/314

Co-authored-by: Raphael <rderbier@gmail.com>